requirements.txt
streamlit
streamlit_lightweight_charts
pandas
numpy
psycopg2-binary
plotly
scipy